fix some bugs

+++ backend/functions/delete.php
@@ -1,5 +1,6 @@
 <?php
 require_once __DIR__ . '/../core/init.php';
+require_once __DIR__ . '/checkUser.php';
 
 $dbh = DB::getInstance();
 
@@ -7,7 +8,7 @@ if(!empty($_POST["token"])) {
   $user_id = Token::check($_POST["token"]);
   if($user_id) {
     // delete todoList 
-    if (!empty($_POST["listId"])) {
+    if (!empty($_POST["listId"]) && empty($_POST["taskId"])) {
       $list_id = $_POST["listId"];
 
       $sql_del_project = "DELETE FROM projects WHERE id = :pid AND user_id = :uid";
@@ -20,15 +21,17 @@ if(!empty($_POST["token"])) {
       $reply = '{"reply" : "TODO List Successfuly DELETED from DB"}';
     }
     // delete TASK
-    elseif (!empty($_POST["taskId"])) {
+    elseif (!empty($_POST["taskId"]) && !empty($_POST["listId"])) {
       $taskId = $_POST["taskId"];
-
+      $list_id = $_POST["listId"];
+      if(checkUser($dbh, $user_id, $list_id, $taskId)){
       $sql = "DELETE FROM tasks WHERE id = :task_id";
-
       $sth = $dbh->exec($sql);
       $sth->execute([':task_id' => $taskId]);
-
       $reply = '{"reply" : "Task Successfuly DELETED from DB"}';
+      } else {
+        $reply = '{ "error" : "Operation not executed. Incoming data not correct" }';
+      }
     }
     else {
       $reply = '{ "error" : "Operation not executed. Incoming data not correct" }';
